+++ Automate/Framework/Models/DataModel.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using System.Runtime.Serialization;
+using Pathoschild.Stardew.Common;
 
 namespace Pathoschild.Stardew.Automate.Framework.Models
 {
@@ -9,12 +11,26 @@ namespace Pathoschild.Stardew.Automate.Framework.Models
         ** Accessors
         *********/
         /// <summary>The name to use for each floor ID.</summary>
-        public Dictionary<int, string> FloorNames { get; set; }
+        public Dictionary<int, DataModelFloor> FloorNames { get; set; } = new();
 
         /// <summary>Mods which add custom machine recipes and require a separate automation component.</summary>
-        public DataModelIntegration[] SuggestedIntegrations { get; set; }
+        public DataModelIntegration[] SuggestedIntegrations { get; set; } = new DataModelIntegration[0];
 
         /// <summary>The configuration for specific machines by ID.</summary>
-        public IDictionary<string, ModConfigMachine> DefaultMachineOverrides { get; set; }
+        public Dictionary<string, ModConfigMachine> DefaultMachineOverrides { get; set; } = new();
+
+
+        /*********
+        ** Public methods
+        *********/
+        /// <summary>Normalize the model after it's deserialized.</summary>
+        /// <param name="context">The deserialization context.</param>
+        [OnDeserialized]
+        public void OnDeserialized(StreamingContext context)
+        {
+            this.FloorNames ??= new();
+            this.SuggestedIntegrations ??= new DataModelIntegration[0];
+            this.DefaultMachineOverrides = this.DefaultMachineOverrides.ToNonNullCaseInsensitive();
+        }
     }
 }