+++ LookupAnything/docs/release-notes.md
@@ -2,7 +2,9 @@
 
 # Release notes
 ## Upcoming release
-Fix error getting harvest dates for potted tea bushes.
+* Added option to show all gift tastes in item/NPC lookups.
+* Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+* Fixed error getting harvest dates for potted tea bushes.
 
 ## 1.35.7
 Released 30 November 2021 for SMAPI 3.13.0 or later.